added create DaiLy

+++ Models/ChiTietPhieuXuat.cs
@@ -5,9 +5,9 @@ namespace QuanLyDaiLy.Models
     public class ChiTietPhieuXuat
     {
         [Key]
-        public long MaChiTietPhieuXuat { get; set; } = 0;
-        public long MaPhieuXuat { get; set; } = 0;
-        public long MaMatHang { get; set; } = 0;
+        public int MaChiTietPhieuXuat { get; set; } = 0;
+        public int MaPhieuXuat { get; set; } = 0;
+        public int MaMatHang { get; set; } = 0;
         public int SoLuongXuat { get; set; } = 0;
         public long DonGia { get; set; } = 0;
         public long ThanhTien { get; set; } = 0;
