+++ Models/LoaiDaiLy.cs
@@ -6,7 +6,7 @@ namespace QuanLyDaiLy.Models
     public class LoaiDaiLy
     {
         [Key]
-        public long MaLoaiDaiLy { get; set; } = 0;
+        public int MaLoaiDaiLy { get; set; } = 0;
         public string TenLoaiDaiLy { get; set; } = "";
         public int NoToiDa { get; set; } = 0;
