+++ Models/PhieuThu.cs
@@ -5,8 +5,8 @@ namespace QuanLyDaiLy.Models
     public class PhieuThu
     {
         [Key]
-        public long MaPhieuThu { get; set; } = 0;
-        public long MaDaiLy { get; set; } = 0;
+        public int MaPhieuThu { get; set; } = 0;
+        public int MaDaiLy { get; set; } = 0;
         public DateTime NgayThuTien { get; set; } = DateTime.Now;
         public long SoTienThu { get; set; } = 0;
