+++ Models/PhieuXuat.cs
@@ -5,8 +5,8 @@ namespace QuanLyDaiLy.Models
     public class PhieuXuat
     {
         [Key]
-        public long MaPhieuXuat { get; set; } = 0;
-        public long MaDaiLy { get; set; } = 0;
+        public int MaPhieuXuat { get; set; } = 0;
+        public int MaDaiLy { get; set; } = 0;
         public DateTime NgayLapPhieu { get; set; } = DateTime.Now;
         public long TongTriGia { get; set; } = 0;
