+++ Models/Quan.cs
@@ -6,7 +6,7 @@ namespace QuanLyDaiLy.Models
     public class Quan
     {
         [Key]
-        public long MaQuan { get; set; } = 0;
+        public int MaQuan { get; set; } = 0;
         public string TenQuan { get; set; } = "";
 
         // Navigation property
