+++ Models/ThamSo.cs
@@ -5,7 +5,7 @@ namespace QuanLyDaiLy.Models
     public class ThamSo
     {
         [Key]
-        public long Id { get; set; } = 0;
+        public int Id { get; set; } = 0;
         public int SoLuongDaiLyToiDa { get; set; } = 0;
         public bool QuyDinhTienThuTienNo { get; set; } = true;
     }
