+++ Services/ILoaiDaiLyService.cs
@@ -4,10 +4,10 @@ namespace QuanLyDaiLy.Services
 {
     public interface ILoaiDaiLyService
     {
-        Task<LoaiDaiLy> GetLoaiDaiLyById(long id);
+        Task<LoaiDaiLy> GetLoaiDaiLyById(int id);
         Task<IEnumerable<LoaiDaiLy>> GetAllLoaiDaiLy();
         Task AddLoaiDaiLy(LoaiDaiLy loaiDaiLy);
         Task UpdateLoaiDaiLy(LoaiDaiLy loaiDaiLy);
-        Task DeleteLoaiDaiLy(long id);
+        Task DeleteLoaiDaiLy(int id);
     }
 }
