+++ Services/IMatHangService.cs
@@ -4,11 +4,11 @@ namespace QuanLyDaiLy.Services
 {
     public interface IMatHangService
     {
-        Task<MatHang> GetMatHangById(long id);
+        Task<MatHang> GetMatHangById(int id);
         Task<IEnumerable<MatHang>> GetAllMatHang();
         Task AddMatHang(MatHang matHang);
         Task UpdateMatHang(MatHang matHang);
-        Task DeleteMatHang(long id);
+        Task DeleteMatHang(int id);
         Task<MatHang> GetMatHangByTenMatHang(string tenMatHang);
     }
 }
