+++ Services/IPhieuThuService.cs
@@ -4,12 +4,12 @@ namespace QuanLyDaiLy.Services
 {
     public interface IPhieuThuService
     {
-        Task<PhieuThu> GetPhieuThuById(long id);
+        Task<PhieuThu> GetPhieuThuById(int id);
         Task<IEnumerable<PhieuThu>> GetAllPhieuThu();
         Task AddPhieuThu(PhieuThu phieuThu);
         Task UpdatePhieuThu(PhieuThu phieuThu);
-        Task DeletePhieuThu(long id);
-        Task<IEnumerable<PhieuThu>> GetPhieuThuByDaiLyId(long maDaiLy);
+        Task DeletePhieuThu(int id);
+        Task<IEnumerable<PhieuThu>> GetPhieuThuByDaiLyId(int maDaiLy);
         Task<IEnumerable<PhieuThu>> GetPhieuThuByDateRange(DateTime startDate, DateTime endDate);
     }
 }
