+++ Services/IPhieuXuatService.cs
@@ -4,12 +4,12 @@ namespace QuanLyDaiLy.Services
 {
     public interface IPhieuXuatService
     {
-        Task<PhieuXuat> GetPhieuXuatById(long id);
+        Task<PhieuXuat> GetPhieuXuatById(int id);
         Task<IEnumerable<PhieuXuat>> GetAllPhieuXuat();
         Task AddPhieuXuat(PhieuXuat phieuXuat);
         Task UpdatePhieuXuat(PhieuXuat phieuXuat);
-        Task DeletePhieuXuat(long id);
-        Task<IEnumerable<PhieuXuat>> GetPhieuXuatByDaiLyId(long maDaiLy);
+        Task DeletePhieuXuat(int id);
+        Task<IEnumerable<PhieuXuat>> GetPhieuXuatByDaiLyId(int maDaiLy);
         Task<IEnumerable<PhieuXuat>> GetPhieuXuatByDateRange(DateTime startDate, DateTime endDate);
     }
 }
